refactor: move users-group endpoind in groups
