+++ src/app.ts
@@ -5,7 +5,6 @@ import path from 'path';
 import YAML from 'yamljs';
 import { usersRouter } from './routes/users';
 import { groupsRouter } from './routes/groups';
-import { usersGroupRouter } from './routes/users-group';
 import { errorHandler } from './middleware/error-handler';
 
 const app = express();
@@ -17,7 +16,6 @@ app.use(express.json());
 app.use('/doc', swaggerUI.serve, swaggerUI.setup(swaggerDocument));
 app.use('/users', usersRouter);
 app.use('/groups', groupsRouter);
-app.use('/users-group', usersGroupRouter);
 
 app.use(errorHandler);
