+++ src/repository/group.ts
@@ -1,5 +1,7 @@
 /* eslint-disable no-unused-vars */
+import { db } from './../database/db';
 import { Group, IGroup, GroupModel } from '../model/group';
+import { UserGroup, UserGroupModel } from '../model/user-group';
 
 export interface IGroupRepository {
     getAll(): Promise<GroupModel[]>,
@@ -7,6 +9,7 @@ export interface IGroupRepository {
     create(group: IGroup): Promise<GroupModel>,
     updateOne(id: string, payload: IGroup): Promise<GroupModel | undefined>,
     removeOne(id: string): Promise<boolean>,
+    addUsers(groupId: string, userIds: string[]): Promise<UserGroupModel[]>,
 }
 export class GroupRepository implements IGroupRepository {
     getAll = () => Group.findAll();
@@ -26,5 +29,10 @@ export class GroupRepository implements IGroupRepository {
     removeOne = async (id: string) => {
         const group = await Group.destroy({ where: { id } });
         return Boolean(group);
+    };
+
+    addUsers = async (groupId: string, userIds: string[]) => {
+        const usersAndGroup = userIds.map((userId) => ({ userId, groupId }));
+        return await db.transaction(async (t) => await UserGroup.bulkCreate(usersAndGroup, { transaction: t }));
     };
 }
