--- src/routes/users-group.ts
@@ -1,6 +1,0 @@
-import { Router } from 'express';
-import usersGroupController from '../controllers/users-group';
-
-export const usersGroupRouter = Router();
-
-usersGroupRouter.post('/', usersGroupController.create);
