+++ src/services/groups.ts
@@ -1,5 +1,6 @@
 /* eslint-disable no-unused-vars */
 import { IGroup } from '../model/group';
+import { IUsersGroup } from '../model/user-group';
 import { IGroupRepository } from '../repository/group';
 
 interface IGroupService {
@@ -8,6 +9,7 @@ interface IGroupService {
     create(group: IGroup): Promise<IGroup>,
     update(id: string, payload: IGroup): Promise<IGroup | undefined>,
     remove(id: string): Promise<boolean>,
+    addUsers(groupId: string, userIds: string[]): Promise<IUsersGroup[]>,
 }
 
 export class GroupService implements IGroupService {
@@ -24,4 +26,6 @@ export class GroupService implements IGroupService {
     update = (id: string, data: IGroup) => this.groupRepository.updateOne(id, data);
 
     remove = (id: string) => this.groupRepository.removeOne(id);
+
+    addUsers = (groupId: string, userIds: string[]) => this.groupRepository.addUsers(groupId, userIds);
 }
